refactor: rename utils package to retry to satisfy linter
- Rename pkg/utils to pkg/retry to avoid 'meaningless package names' lint error
- Rename RetryConfig to Config to avoid stuttering (retry.Config vs retry.RetryConfig)
- Update imports in controller_nfs.go and controller_nvmeof.go
